Fixed the refcounting for realtime queues.
Here's the "issue" at the moment. The current app_queue procedure makes it so that a realtime queue will be destroyed
the first time that a call is attempted on the queue once it is removed from realtime. This is a bit unnatural and also
means the queue stays in memory longer than its actual life. The reason for this is that realtime queues have no way to
be marked "dead" until we search for them in realtime, and the only way we'll search for them in realtime is a caller attempts
to enter the queue (or if queue show or some other cli or manager command needs to find the queue)