The main purpose of this patch is to remove the leftover calls to get_video_frames() in chan_oss and chan_alsa, since video input is now handled entirely by the video thread. At the same time, change the period of the thread from 1s to 50ms.
This might fix the ast_queue corruption that i and Matteo were
seeing, as the two threads were contending for the video device
and the codec.

The second (and largest, but comment-only) part of the patch is to
start documenting the way we add video support, and list the various
video representations that are in use.  As you can see from the
comment we have many different representations to handle (local_source,
encoder_in, encoder_out, local_display, decoder_in, decoder_out,
remote_display).

The current implementation completely overlooked this, assuming that
the geometry is the same for all representations, and also that
bitstream formats are the same for some of these representations.
Now that I know, i must patiently add descriptors for all of them,
and possibly invoke img_convert/sws_scale as appropriate.