Make some notes about a problem I found with the OPTIONs handler while working with the bug tracker. Since we don't authenticate devices (peers/users) on OPTIONS we don't have the proper context set for the user/peer.
However, we might not want to process an authentication for every OPTIONS, so we could
have a config option for this, "optionsforceok" to always answer 200 OK on the request
and not check device or destination, nor add a SDP. If Asterisk sends the OPTIONs request,
it doesn't care about the reply. Some devices use OPTIONs to discover capabilities,
since we should answer like an INVITE from the device and we need to support that properly
too, which we don't today.

So much to do :-)