Branch to try to track down 491 handling on Asterisk 1.4 reinvite races.
Asterisk 1.4 actually denies an INVITE if we have an outstanding
INVITE in the same call. Asterisk 1.2 happily accepts anything 
regardless.

This causes a problem where the 491 on a reinvite is handled 
as an error. We need to remember that there's a re-invite
and not an invite that we're dealing with and postpone it.

Related to bug #10481