Several changes
1. I changed the "rules" list to have a lock, and use this list lock instead of the queue's lock since it is much more logical
2. Laureano on IRC pointed out a potential memory leak which is fixed here. Thanks Laureano! :)
3. I realized that I was completely mishandling relative penalty changes in the insert_penaltychange function. It's fixed now.
4. I now initialize the list of rules in the init_queue function.
5. I added the logic to change values of the QUEUE_MAX_PENALTY in wait_our_turn and queue_exec.

At this point, you could call what's written a really rough draft. It SHOULD work as long as you
don't make any changes to the penalty rules while someone is in the queue (In other words no reloads or realtime updates)

My next step is to test what's written to see if things are actually going as planned. After that I will work on 
refining the code which has been written and try to integrate other enhancements.