Creating this branch as a means of fixing issue #11498. The change necessary is somewhat large, and before all is said and done, it may be too big a change for 1.4. Therefore, I am making this branch against trunk and will backport the change to 1.4 if it makes sense to do so.
There are two issues here regarding the queue log. It is possible for certain RINGNOANSWER
log entries to not have a member name listed. Also, when using the ringall strategy, if no one
answers, then the RINGNOANSWER entry generated will randomly have one of the member's names listed.
While this may not sound like a huge deal, it has exposed a need for app_queue to change the way in
which it manages which members it is calling.