Bring in the code from team/russell/jack/.
Add a new module, app_jack, which provides interfaces to JACK, the Jack
Audio Connection Kit (http://www.jackaudio.org/).  Two interfaces are
provided; there is a JACK() application, and a JACK_HOOK() function.  Both
interfaces create an input and output JACK port.  The application makes
these ports the endpoint of the call.  The audio coming from the channel
goes out the output port and whatever comes back in on the input port is
what gets sent to the channel.  The JACK_HOOK() function turns on a JACK
audiohook on the channel.  This lets you run the audio coming from a
channel through JACK, and whatever comes back in is what gets forwarded
on as the channel's audio.  This is very useful for building custom
vocoders or doing recording or analysis of the channel's audio in another
application.

In case anyone is curious, the platform that inspired me to write this is
PureData (http://puredata.info/).  I wrote these JACK interfaces so that I
could use Pd to do interesting things with the audio of phone calls ...