The bug being addressed here is related to issue #11698.
The issue is that it is possible for a translation frame to still be in use
when the translation path gets destroyed.  Specifically, this can happen if
a masquerade happens in between a call to ast_read() and ast_write(), since
the channel is not locked between these two operations.

This hack of a fix is written in such a way that it is API and ABI compatible
with previous versions of Asterisk 1.4.

First, the "has_timing_info" field in ast_frame was change to be a flags field.
A flag was added that gets set when a translator outputs an internal frame.
Then, when a frame gets free'd, if it has this flag set, a hint is given to
the translation core.  So, if the translator was requested to be destroyed while
this frame was still in use, it can be destroyed once the frame is sent to
ast_frame_free(), meaning that it is no longer needed.