Add AGIstate to manager events NewExten and CoreShowChannels
This is to tell manager controllers that these events are not generated from the dialplan, but from an AGI control script.

We might want to add a manager event when the AGI script releases control too, so we can follow the sequence of events.
AGI starts with newevent with an AGI application, then runs until there's no more AGIstates reported (or the new
end-of-agi-control manager event that only exists in my imagination yet).