Merged revisions 110831 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r110831 | mmichelson | 2008-03-25 19:02:31 -0500 (Tue, 25 Mar 2008) | 6 lines

This ensures that the manager interface is not enabled by default. Prior to this
change, it was possible to start Asterisk with the manager interface enabled, then
either comment out the enabled option or make manager.conf unopenable and the manager
interface would still be enabled.


........