Save some code that I wrote on a plane back to Huntsville today.
We have had multiple discussions in the past of a "data retrieval" API for
Asterisk to accomplish the following things, mainly:
  1) Provide a more centralized implementation for the existing interfaces into
     Asterisk to retrieve data (CLI, Manager Interface, SNMP, etc.).
  2) By using a more common implementation, ensure that all interfaces into
     Asterisk have access to all of the same data.
  3) Provide an easier way to retrieve data between parts of Asterisk, (like
     actually making it possible for the SNMP module to access information
	 on SIP peers, without having to do something insane like connect to the
	 local manager interface!)

I have made a first pass at something to accomplish these goals.  The main API
definition and implementation are pretty simple, I think, and exist in main/data.c
and include/asterisk/data.h.

The code that demonstrates and tests what I have so far is:
  - In channels/chan_sip.c, I have implemented a single "data provider", called
    "sip_settings".  The "sip_settings" data provider would eventually provide
	all of what the "sip show settings" CLI command currently provides, but through
	this new interface.
  - There is an example of accessing this data in the CLI command that currently
    exists in main/data.c.  There is a "data get" CLI command which allows you to
	retrieve data from a data provider and have it presented to you at the CLI.
	So, "data get sip_settings" currently provides some information.

So, in theory, once this interface, or something like it gets agreed upon, a very
large number of the existing CLI commands and manager actions would be re-written
to this new interface.  Once they are on the new interface, they would then be
magically available to the CLI, manager interface, SNMP, or whatever other code
that someone wanted to write to access this information.