Blocked revisions 112207 via svnmerge
........
r112207 | jpeeler | 2008-04-01 12:53:08 -0500 (Tue, 01 Apr 2008) | 3 lines

This adds DNS SRV record support to DNS manager.  If there is a SRV record for a given domain, the hostname and port listed in the SRV record will be used. If no SRV record exists or a SRV lookup is not attempted, the DNS lookup on the specified domain will be performed as normal. Chan_sip has been modified to take advantage of the new SRV support.


........