This is a progress checkin. I suspect this doesn't even actually compile at this point...
Now at module load time, a configuration file, subscribers.conf, is parsed to see what users
are defined and what their filter preferences are. Right now, the only way of specifying filtering
info is to have an "event=" line in the config file. This will correspond to an event type, meaning
that subscriber wishes to see output when an event of that type occurs. In order to accommodate this,
I added a new mapping in event.c which maps enum ast_event_types to a string representation of their names
as well as a function to get an ast_event_type given its name as input.

This checkin covers the setup of the filter, and leaves the functions in a state where they should
be easy to expand on later, since they will invariably need lots of expansion. This does not yet implement
the use of the filters when subscribing to events. That will be the next item on the agenda once I have
what's written compiling properly.